refactor(env): split env.ts into env.client.ts and env.server.ts
- Isolates client env module from server env side effects
- Importing clientEnv no longer executes serverEnvSchema.parse()
- Prevents future required server vars from crashing client bundles
- Updates imports in client.ts, server.ts, proxy.ts
