+++ lib/env.client.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 
 // ---------------------------------------------------------------------------
-// Why two separate env objects?
+// Why literal process.env references?
 //
 // Next.js inlines NEXT_PUBLIC_* variables into the browser bundle at BUILD
 // time, but only when they appear as *literal* expressions like:
@@ -14,12 +14,9 @@ import { z } from "zod";
 //
 //   envSchema.parse(process.env)                 // breaks in the browser
 //
-// So we split the config into:
-//   - clientEnv  — uses literal references, safe in both browser and server
-//   - serverEnv  — uses process.env object, server-side only
+// So each value below MUST be a literal `process.env.NEXT_PUBLIC_*` reference
+// so Next.js can find and inline it during the build.
 // ---------------------------------------------------------------------------
-
-// --- Client env (safe in browser + server) --------------------------------
 
 const clientEnvSchema = z.object({
   NEXT_PUBLIC_SUPABASE_URL: z
@@ -36,20 +33,8 @@ const clientEnvSchema = z.object({
     ),
 });
 
-// Each value MUST be a literal `process.env.NEXT_PUBLIC_*` reference so
-// Next.js can find and inline it during the build.
 export const clientEnv = clientEnvSchema.parse({
   NEXT_PUBLIC_SUPABASE_URL: process.env.NEXT_PUBLIC_SUPABASE_URL,
   NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY:
     process.env.NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY,
 });
-
-// --- Server env (server-side only) ----------------------------------------
-
-const serverEnvSchema = z.object({
-  SUPABASE_SECRET_KEY: z.string().optional(),
-});
-
-// Passing the full process.env object is fine here because server-side code
-// always has access to all environment variables at runtime.
-export const serverEnv = serverEnvSchema.parse(process.env);
